Completed the Login Page
Today I did the HTML and CSS code for the Login component of my PWA. I incorporated my logo, a title, and container where users are able to input their usernames and passwords. Small icons and a background image is incorporated for aesthetics, as well as a 'sign in' button to transfer users to the site. This will be achieved later on.