Began the Browsing Page
Through JavaScript, I created connections between the login and browsing pages, where a valid user and password must be inserted to sign in. If the incorrect data in written, a small note will appear on the page. If successful, the user is transported to the browsing page. I completed the navigation bar for this section as well as the search bar, while being inbetween the TV show icons.