Minor alterations and added details
I decided to make some design changes due to my ideas changing for the PWA. I deleted the search bar, and decided to include fewer pages to ensure plenty of attention could be placed onto those. I also solved the issue with the genre not appearing in the pop-up, which was due to the data-genre. Then, I made a button inside the pop-ups which I will code to transfer to the form page where the database will be formed.