Started on the next page
I completed making the buttons on the browse page, which allowed users to transfer to the 'showcase' form page upon selecting a TV show they've completed watching. I also completed making the form, with input for the title, genre, rating and review. Through javascript, the final review is displayed in the next 'watched' page, where readers can look at all their watched shows, edit, and delete them if desired. The styling for the 'watched' page still needs to be done, as well as slight alterations to the form.