Completed all three pages
I have finally completed the three pages of my website. I first altered the genre component in the form to make it a drop-down selection rather than scroll, as it crafted a cleaner look. I then altered the function for the rating system in Java, as even though it was supposed to only allow numbers 0-5, it allowed any integers. This was fixed with an error message also appearing if users were to enter unwanted numbers. In addition, I transformed my 'watched' page by changing up the aesthetics with cleaner grids and colours. I also enhanced the 'update' button, as when users are navigated back to the form, the previous information from their review before remains in the inputs which they can then change. It also doesn't appear as doubles anymore after being submitted. I incorporated a title to the Browse page, and responsive adjustments for when the screen size changes.