Backend Work and Added More Shows
I made the database file for my backend and decided to insert a few more shows in the browsing category. I also altered the headings of some of the pages as Miss suggested. There were many issues and confusion I encountered at this stage, so there many edits I made to the code, going back and forth between pieces to see if I was doing everything right. I also followed the Google Doc when making certain features of the backend such as the database.