Front and Backend Connection attempts
First I made the filter for the frontend in the Watched component, the html was edited to insert all the possible genres, the javascript allowed for only those specific logs to appear and the css styled the dropdown. I also attempted changing the server.js as well as the watched.js multiple times to get the back and front end to connect, but it is not working yet.