Ensuring Everything Works
I made some edits to my database by adding, updating and deleting for the video component of the task. The installation also stopped working for a bit, but then I realised the issue was with the urls I made in the service worker and manifest files, so I changed all the 'showcase' statements to 'myPWA' instead and it began working successfully.