Software Second Task - Making a Secure Login Page
I completely altered my login page for this task as the key component is to enhance security. I first incorporated both a login and registration page through html, css and java, then crafted animations, password validation with rules, CAPTCHA and a limit to the number of times a user can attempt their login, with a timed blocking if there's three failed attempts.