Minor Alterations to Appearance
I made some minor alterations to the appearance of my frontend, since the main focus of this task is the security of the login and registration. I first altered the border radius of my photo cards, and made some font, colour and sizing adjustments in the browse and showcase pages. I also altered the error messaging and restrictions on the rating system by ensuring it restricts numbers that aren't between 0-5.