Incorporated Password Validation, Sanitisation and SQL Prevention
I incorporated password validation in the registration page by creating error messages for if the password doesn't fulfil all requirements. In addition to that, users are unable to type in spaces in the login page with the sanitisation, hence decreasing possibilities of threats. Furthermore, if SQL injections are pasted into any input text box, the item is altered in text to display [disallowed input] through sanitisation.