Sanitisation Incorporated Properly in the Review Forms
I realised I had incorporated sanitisation incorrectly in the review form, as it wouldn't turn the SQL injections into the altered text as I desired. Hence I created a component that replaces certain patterns or characters with the text '[disallowed input]' for enhanced security. I also permitted spaces within the form, as it is required for a high UX and functionality when writing reviews.